requirements.txt
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
decorator==5.1.1
EbookLib==0.17.1
executing==0.8.3
idna==3.3
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
lxml==4.8.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-pptx==0.6.21
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
typing-extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
XlsxWriter==3.0.3
